Clean things up
This patch just includes a bunch of cleanup and some changes to logging.
Mostly just running testsa and flushing out any stray issues.

Slight bit of work still to be done on publishing and reporting, that's next
but then we should be on par with sos-ci and ready to add in other components.